refactor: Add ClockifyController
